refactoring mock data
